backend/requirements.txt
asgiref==3.9.1
dataclasses==0.6
Django==5.2.6
django-mongodb-backend==5.2.1
djangorestframework==3.16.1
django-cors-headers==4.3.1
dnspython==2.8.0
pymongo==4.15.0
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
